Refactor bookworm component
